requirements.txt
caldav>=1.3.0
icalendar>=5.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.100.0
python-dateutil>=2.8.2
pytz>=2023.3
